+++ CHANGELOG.md
@@ -14,10 +14,14 @@
     - `OnGuildAvailable` and `OnGuildUnavailable` were merged into `OnGuildCreate` and `OnGuildDelete`. To see whether the guild availability status changed, use the new `BecameAvailable` and `Unavailable` properties in the event args.
     - `OnDMChannel*` events were removed (bots no longer receive events for DM channels as of Gateway v8).
     - `OnChannel*` events now emit a `DiscordChannel` instead of a `DiscordGuildChannel` (matches the actual Gateway spec).
+    - `UpdateStatusAsync` was renamed to `UpdatePresenceAsync`.
 - `Shard.StartAsync` now requires either a `GatewayIntent` or `ShardStartConfig` argument.
 - Renamed `DiscordGame` to `DiscordActivity` (to match Gateway v8 changes).
 - Renamed `DiscordGameType` to `DiscordActivityType` (to match Gateway v8 changes).
 - Removed `DiscordUserPresence.Game` (use `Activities.FirstOrDefault()` instead).
+- Renamed `StatusOptions` to `PresenceOptions`.
+    - `PresenceOptions` now takes a list of activities instead of a single `Game`.
+- Renamed `GameOptions` to `ActivityOptions`.
 - `DiscordVoiceConnection.SetSpeakingAsync` will now throw an `InvalidOperationException` if called before being fully connected.
 - `DiscordVoiceConnection.ClearVoiceBuffer` will now throw an `InvalidOperationException` if called before being fully connected.
 - `DiscordVoiceConnection.ConnectAsync` no longer checks if the application is allowed to join the voice channel. If the application is not allowed to join the connection will still fail, but it will be attempted.